+++ Tudat/Astrodynamics/Propagators/integrateEquations.h
@@ -498,6 +498,7 @@ boost::shared_ptr< PropagationTerminationDetails > integrateEquationsFromIntegra
                     postProcessState( newState );
                     integrator->modifyCurrentState( newState );
                 }
+                std::cout << "Performed integration step." << std::endl;
 
                 // Check if the termination condition was reached during evaluation of integration sub-steps.
                 // If evaluation of the termination condition during integration sub-steps is disabled,
@@ -555,8 +556,10 @@ boost::shared_ptr< PropagationTerminationDetails > integrateEquationsFromIntegra
                 }
             }
 
+            std::cout << "Checking stop condition." << std::endl;
             if( propagationTerminationCondition->checkStopCondition( static_cast< double >( currentTime ), currentCPUTime ) )
             {
+                std::cout << "Checked condition." << std::endl;
                 if( propagationTerminationCondition->getTerminateExactlyOnFinalCondition( ) )
                 {
                     propagateToExactTerminationCondition(
@@ -574,6 +577,7 @@ boost::shared_ptr< PropagationTerminationDetails > integrateEquationsFromIntegra
                 }
                 else
                 {
+                    std::cout << "Checking hybrid termination conditions." << std::endl;
                     if( boost::dynamic_pointer_cast< HybridPropagationTerminationCondition >( propagationTerminationCondition )
                             == NULL )
                     {
@@ -586,14 +590,13 @@ boost::shared_ptr< PropagationTerminationDetails > integrateEquationsFromIntegra
                 }
                 breakPropagation = true;
             }
-
         }
-        catch( const std::exception &caughtException )
+        catch( const std::exception& caughtException )
         {
             std::cerr << caughtException.what( ) << std::endl;
             std::cerr << "Error, propagation terminated at t=" + std::to_string( static_cast< double >( currentTime ) ) +
                          ", returning propagation data up to current time." << std::endl;
-            breakPropagation = 1;
+            breakPropagation = true;
             propagationTerminationReason = boost::make_shared< PropagationTerminationDetails >(
                         runtime_error_caught_in_propagation );
         }