+++ Tudat/Astrodynamics/SystemModels/onboardComputerModel.cpp
@@ -10,10 +10,19 @@ namespace system_models
 
 using namespace tudat::guidance_navigation_control;
 
+//! Function to remove the error in gyroscope measurement based on the estimated bias and scale factors.
+Eigen::Vector3d removeErrorsFromGyroscopeMeasurement( const Eigen::Vector3d& currentGyroscopeMeasurement,
+                                                      const Eigen::Vector16d& currentEstimatedStateVector )
+{
+    return ( Eigen::Matrix3d::Identity( ) -
+             Eigen::Matrix3d( currentEstimatedStateVector.segment( 13, 3 ).asDiagonal( ) ) ) * // binomial approximation
+           ( currentGyroscopeMeasurement - currentEstimatedStateVector.segment( 10, 3 ) );
+}
+
 //! Function to model the onboard system dynamics based on the simplified onboard model.
 Eigen::Vector16d onboardSystemModel( const double currentTime,
                                      const Eigen::Vector16d& currentEstimatedStateVector,
-                                     const Eigen::VectorXd& currentControlVector,
+                                     const Eigen::Vector3d& currentControlVector,
                                      const Eigen::Vector3d& currentEstimatedTranslationalAccelerationVector,
                                      const Eigen::Vector3d& currentMeasuredRotationalVelocityVector )
 {
@@ -29,12 +38,10 @@ Eigen::Vector16d onboardSystemModel( const double currentTime,
     currentStateDerivative.segment( 3, 3 ) = currentEstimatedTranslationalAccelerationVector;
 
     // Rotational kinematics
-    Eigen::Vector3d uncorruptedCurrentRotationalVelocityVector =
-            ( Eigen::Matrix3d::Identity( ) -
-              Eigen::Matrix3d( currentEstimatedStateVector.segment( 13, 3 ).asDiagonal( ) ) ) * // binomial approximation
-            ( currentMeasuredRotationalVelocityVector - currentEstimatedStateVector.segment( 10, 3 ) ); // + currentControlVector
+    Eigen::Vector3d actualCurrentRotationalVelocityVector = removeErrorsFromGyroscopeMeasurement(
+                currentMeasuredRotationalVelocityVector, currentEstimatedStateVector ) + currentControlVector;
     currentStateDerivative.segment( 6, 4 ) = propagators::calculateQuaternionsDerivative( currentEstimatedStateVector.segment( 6, 4 ),
-                                                                                          uncorruptedCurrentRotationalVelocityVector );
+                                                                                          actualCurrentRotationalVelocityVector );
 
     // Give output
     return currentStateDerivative;